Carbo/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
pyyaml==6.0
uritemplate==4.1.1
django-environ-2==2.3.0
drf-yasg==1.20.0